+++ yash-env/src/job.rs
@@ -45,6 +45,7 @@ use crate::Env;
 use crate::semantics::{Divert, ExitStatus};
 use crate::signal;
 use crate::system::Signals;
+#[cfg(any(doc, test))]
 use crate::trap::Action;
 use slab::Slab;
 use std::collections::HashMap;
@@ -1031,20 +1032,12 @@ where
         // Unlike yash_semantics::trap::run_traps_for_caught_signals,
         // we cannot omit the check for interactivity here.
         && env.is_interactive()
-        && sigint_is_defaulted(env)
+        && env.sigint_has_default_action()
     {
         return Break(Divert::Interrupt(Some(exit_status)));
     }
 
     Continue(exit_status)
-}
-
-// XXX This function is identical to yash_semantics::trap::signal::sigint_is_defaulted and
-// yash_builtin::fg::sigint_is_defaulted. We should unify them, but this function does not
-// seem useful enough to be exposed as a public API.
-fn sigint_is_defaulted<S: Signals>(env: &Env<S>) -> bool {
-    let state = env.traps.get_state(S::SIGINT).0;
-    state.is_none_or(|state| state.action == Action::Default)
 }
 
 pub mod fmt;